% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_steiner_tree.R
\name{cut_steiner_tree}
\alias{cut_steiner_tree}
\title{divide the Steiner tree}
\usage{
cut_steiner_tree(st_res, k)
}
\arguments{
\item{st_res}{a list contains steiner tree and plot (result of PlotSteinertree)}

\item{k}{number of groups to divide}
}
\value{

}
\description{
divide the Steiner tree
}
