% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Markers.R
\name{Find_Markers}
\alias{Find_Markers}
\title{Find expression markers for all identity classes}
\usage{
Find_Markers(object, assay = NULL, binarizeMethod = "median",
  FoldChange = 2, p.adj = 0.05, cores = 1)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{assay of the object, default is the default assay}

\item{binarizeMethod}{Either "median" (default) or "naive" or a numeric cutoff}

\item{FoldChange}{default is 2}

\item{p.adj}{default is 0.05}

\item{cores}{default is 1}
}
\description{
Find expression markers for all identity classes
}
