% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enrich_plot.R
\name{get_enrich_plot}
\alias{get_enrich_plot}
\title{GO enrichment of Steiner tree}
\usage{
get_enrich_plot(st_res, group)
}
\arguments{
\item{st_res}{result of cut_steiner_tree (or you can add group information manually)}

\item{group}{the group of genes you want to do GO enrichment}
}
\value{
a list with enrich_res and plot
}
\description{
GO enrichment of specified group in Steiner tree
}
