% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_steiner_plot.R
\name{get_steiner_plot}
\alias{get_steiner_plot}
\title{Title}
\usage{
get_steiner_plot(ig_obj, label_num = 20, name,
  typeof_node_size = c("pagerank", "degree"), only_label_terminal = T,
  terminals, marker_genes = NULL, module_genes = NULL, nodes = NULL)
}
\arguments{
\item{ig_obj}{the Steiner tree, an igraph object}

\item{label_num}{number of nodes you want to label, default is 20}

\item{name}{name of the plot}

\item{typeof_node_size}{'pagerank' or 'degree', former is default}

\item{only_label_terminal}{default is F}

\item{terminals}{terminals of the Steiner tree}

\item{marker_genes}{marker_genes of the Steiner tree}

\item{module_genes}{module_genes of the Steiner tree}

\item{nodes}{the nodes you want to labeled in the Steiner tree}
}
\value{

}
\description{
Title
}
