% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPI_Biogrid.R
\name{getPPI_Biogrid}
\alias{getPPI_Biogrid}
\title{get PPI from Biogrid}
\usage{
getPPI_Biogrid(object = NULL, species = 9606)
}
\arguments{
\item{species}{NCBI taxonomy identifier of the species to query for homologs, default is 9606(Homosapiens)}
}
\value{
PPI matrix
}
\description{
get PPI from Biogrid
}
