% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPI_String.R
\name{getPPI_String}
\alias{getPPI_String}
\title{get PPI from String}
\usage{
getPPI_String(object, version = "10", species = 9606,
  score_threshold = 600, save = FALSE)
}
\arguments{
\item{object}{an Seurat object which contains the expression matrix}

\item{version}{String version}

\item{species}{NCBI taxonomy identifier of the species to query for homologs, default is 9606(Homosapiens)}

\item{score_threshold}{threshold on the score of the interaction, default is 600}

\item{save}{whether save the result as a Rdata file, default is FALSE}
}
\value{
PPI matrix
}
\description{
get PPI from String
}
