% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2seurat.R
\name{mat2seurat}
\alias{mat2seurat}
\title{processing function}
\usage{
mat2seurat(
  Data,
  min_nFeature = 1000,
  min_nCount = 10000,
  features = NULL,
  selection.method = c("mean.var.plot", "vst", "dispersion"),
  nfeatures = NULL,
  mean.cutoff = c(0.1, Inf),
  dispersion.cutoff = c(0.1, Inf),
  names.filed = 1,
  names.delim = "_"
)
}
\arguments{
\item{Data}{gene*cell matrix}

\item{min_nFeature}{Include cells where at least this many features are detected}

\item{min_nCount}{Include cells where at least this many counts are detected}

\item{features}{Vector of features to use as variable features.}

\item{selection.method}{How to choose top variable features.}

\item{nfeatures}{Number of features to select as top variable features; only used when selection.method is set to 'dispersion' or 'vst'}

\item{mean.cutoff}{A two-length numeric vector with low- and high-cutoffs for feature means}

\item{dispersion.cutoff}{A two-length numeric vector with low- and high-cutoffs for feature dispersions}
}
\value{
Seurat object
}
\description{
transfer raw matrix data to Seurat object
}
