% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.SCORE.R
\name{R.SCORE}
\alias{R.SCORE}
\title{Title}
\usage{
R.SCORE(Data, PPI = "Biogrid", species = 9606, score_threshold = 600,
  metric = c("cor", "rho", "phi", "phs"), module_min = 3,
  max_step = 5, AUCRank = 400)
}
\arguments{
\item{Data}{gene*cell matrix or a seurat object.}

\item{PPI}{protein-protein network, you can provide a matrix if you have PPI data,
or you can let PPI = 'Biogrid' or 'String', the system will download corresponding PPI data.}

\item{species}{NCBI taxonomy identifier of the species to query for homologs, default is 9606(Homosapiens),
only needed when PPI = 'Biogrid' or 'String'.}

\item{score_threshold}{threshold on the score of the interaction, default is 600, only needed when PPI = 'String'}

\item{metric}{A character string. The metric used to calculate association.
Choose one of 'cor', 'rho', 'phi', 'phs'}

\item{module_min}{Integer number. The minimum size of a module to be accepted.}

\item{max_step}{Integer number. The maximum step run in the community detect.}

\item{AUCRank}{Integer number. The number of highly-expressed genes to include when computing AUCell.}
}
\value{
a seurat object
}
\description{
Title
}
