% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steinertree.R
\name{steinertree}
\alias{steinertree}
\title{get steiner tree}
\usage{
steinertree(terminals, graph, method = "sp", color = T, weighted = T,
  optimize = T)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
calculate the steiner tree, two methods can be chosen
}
