% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPI_Biogrid.R
\name{getPPI_Biogrid}
\alias{getPPI_Biogrid}
\title{get PPI from Biogrid}
\usage{
getPPI_Biogrid(
  object = NULL,
  version = "3.5.174",
  species = 9606,
  save = FALSE
)
}
\arguments{
\item{version}{Version of BioGRID, such as '3.5.174'}

\item{species}{NCBI taxonomy identifier of the species to query for homologs, default is 9606(Homosapiens)}

\item{save}{whether save the result as a Rdata file, default is FALSE}
}
\value{
PPI matrix
}
\description{
get PPI from Biogrid
}
