% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSteinertree.R
\name{PlotSteinertree}
\alias{PlotSteinertree}
\title{Title}
\usage{
PlotSteinertree(object, ident = NULL, weighted = T, onlysteiner = T)
}
\arguments{
\item{ident}{which class you want to show, default: all}

\item{weighted}{if it is TRUE, the network will be weighted, default TRUE}

\item{onlysteiner}{if it is TRUE, will only show the steiner tree, default TRUE}
}
\description{
Title
}
