% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.SCORE.R
\name{R.SCORE}
\alias{R.SCORE}
\title{R.SCORE}
\usage{
R.SCORE(
  Data,
  PPI = "Biogrid",
  species = 9606,
  score_threshold = 600,
  metric = c("cor", "rho", "phi", "phs"),
  module_min = 3,
  max_step = 10,
  AUCRank = 400,
  SeqMethod = "10X",
  nCores = 1
)
}
\arguments{
\item{Data}{gene*cell matrix or a seurat object.}

\item{PPI}{protein-protein network, you can provide a matrix if you have PPI data,
or you can let PPI = 'Biogrid' or 'String', the system will download corresponding PPI data.}

\item{species}{NCBI taxonomy identifier of the species to query for homologs, default is 9606(Homosapiens),
only needed when PPI = 'Biogrid' or 'String'.}

\item{score_threshold}{threshold on the score of the interaction, default is 600, only needed when PPI = 'String'}

\item{metric}{A character string. The metric used to calculate association.
Choose one of 'cor', 'rho', 'phi', 'phs'}

\item{module_min}{Integer number. The minimum size of a module to be accepted.}

\item{max_step}{Integer number. The maximum step run in the community detect.}

\item{AUCRank}{Integer number. The number of highly-expressed genes to include when computing AUCell.}

\item{SeqMethod}{A charater string. Sequening methods, default is '10X'}

\item{nCores}{Number of cores to use for computation.}
}
\value{
a seurat object
}
\description{
explore single-cell RNA-seq data with the view of molecular networks.
}
